Update main.js
Instead of having the resultButton do two functions seperately, i made it activate a function called functionTrain. 

functionTrain will execute resultsPop and use the value of the #amount input and compare it with the intern local variable inside the for loop. 

Essentialy the for loop will stop if i (the local var) is smaller or equal to the number we gave in the #amount input. Inside the for loop we put Results(); that will produce mutliple numbers.